refactor: make dist and src consistent for usage and update tests
