+++ src/providers/README.md
@@ -1,0 +1,1 @@
+Module should not depend on runtime (any directory outside)
