+++ src/providers/auth0.ts
@@ -1,4 +1,4 @@
-import { assignDefaults } from '../utils'
+import { assignDefaults } from './_utils'
 
 export function auth0 (_nuxt, strategy) {
   assignDefaults(strategy, {
