+++ templates/plugin.js
