+++ test/fixture/pages/index.vue
