+++ test/fixture/pages/secure.vue
