+++ test/fixture/plugins/auth.js
