+++ test/fixture/store/index.js
